pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "audio-intent-pipeline"
version = "1.0.0"
description = "Production-aware AI voice pipeline: STT → LLM intent classification → structured output"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "structlog>=24.2.0",
    "httpx>=0.27.0",
    "deepgram-sdk==6.0.1",
    "openai>=1.35.0",
    "python-multipart>=0.0.9",
    "pydub>=0.25.1",
    "python-magic>=0.4.27",
    "openai-whisper>=20231117",
    "python-dotenv>=1.0.1",
    "anyio>=4.4.0",
    "audioop-lts>=0.2.2",
    "google-genai==1.65.0",
    "streamlit>=1.54.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-mock>=3.14.0",
    "ruff>=0.5.0",
    "black>=24.4.2",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "anyio>=4.12.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]
